requirements.txt
parakeet-mlx>=0.1.0
mlx-lm>=0.18.0
mlx-whisper>=0.4.0
sounddevice>=0.4.0
scipy>=1.11.0
pynput>=1.7.0
pyperclip>=1.8.0
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
